+++ auto_tests/conference_test.c
@@ -88,7 +88,7 @@ static void handle_conference_connected(
     state->invited_next = true;
 }
 
-static uint16_t num_recv;
+static uint32_t num_recv;
 
 static void handle_conference_message(
     Tox *tox, uint32_t groupnumber, uint32_t peernumber, Tox_Message_Type type,
@@ -132,12 +132,12 @@ static void disconnect_toxes(uint32_t tox_count, Tox **toxes, State *state,
     bool invert = false;
 
     do {
-        for (uint16_t i = 0; i < tox_count; ++i) {
+        for (uint32_t i = 0; i < tox_count; ++i) {
             disconnect_now[i] = exclude[i] || (invert ^ disconnect[i]);
         }
 
         do {
-            for (uint16_t i = 0; i < tox_count; ++i) {
+            for (uint32_t i = 0; i < tox_count; ++i) {
                 if (!disconnect_now[i]) {
                     tox_iterate(toxes[i], &state[i]);
                     state[i].clock += 1000;
@@ -164,8 +164,8 @@ static bool all_connected_to_group(uint32_t tox_count, Tox **toxes)
 
 static bool names_propagated(uint32_t tox_count, Tox **toxes, State *state)
 {
-    for (uint16_t i = 0; i < tox_count; ++i) {
-        for (uint16_t j = 0; j < tox_count; ++j) {
+    for (uint32_t i = 0; i < tox_count; ++i) {
+        for (uint32_t j = 0; j < tox_count; ++j) {
             const size_t len = tox_conference_peer_get_name_size(toxes[i], 0, j, nullptr);
 
             if (len != NAMELEN) {
@@ -205,7 +205,7 @@ static void run_conference_tests(Tox **toxes, State *state)
 
     ck_assert(NUM_DISCONNECT < NUM_GROUP_TOX);
 
-    for (uint16_t i = 0; i < NUM_DISCONNECT; ++i) {
+    for (uint32_t i = 0; i < NUM_DISCONNECT; ++i) {
         uint32_t disconnect = random_false_index(disconnected, NUM_GROUP_TOX);
         disconnected[disconnect] = true;
 
@@ -220,7 +220,7 @@ static void run_conference_tests(Tox **toxes, State *state)
     uint8_t *save[NUM_GROUP_TOX];
     size_t save_size[NUM_GROUP_TOX];
 
-    for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
+    for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
         if (restarting[i]) {
             save_size[i] = tox_get_savedata_size(toxes[i]);
             ck_assert_msg(save_size[i] != 0, "save is invalid size %u", (unsigned)save_size[i]);
@@ -233,7 +233,7 @@ static void run_conference_tests(Tox **toxes, State *state)
 
     disconnect_toxes(NUM_GROUP_TOX, toxes, state, disconnected, restarting);
 
-    for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
+    for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
         if (restarting[i]) {
             struct Tox_Options *const options = tox_options_new(nullptr);
             tox_options_set_savedata_type(options, TOX_SAVEDATA_TYPE_TOX_SAVE);
@@ -249,7 +249,7 @@ static void run_conference_tests(Tox **toxes, State *state)
     if (check_name_change_propagation) {
         printf("changing names\n");
 
-        for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
+        for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
             char name[NAMELEN + 1];
             snprintf(name, NAMELEN + 1, NEW_NAME_FORMAT_STR, state[i].index);
             tox_self_set_name(toxes[i], (const uint8_t *)name, NAMELEN, nullptr);
@@ -264,7 +264,7 @@ static void run_conference_tests(Tox **toxes, State *state)
 
     printf("running conference tests\n");
 
-    for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
+    for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
         tox_callback_conference_message(toxes[i], &handle_conference_message);
 
         iterate_all_wait(NUM_GROUP_TOX, toxes, state, ITERATION_INTERVAL);
@@ -286,8 +286,8 @@ static void run_conference_tests(Tox **toxes, State *state)
     ck_assert_msg(num_recv == NUM_GROUP_TOX, "failed to recv group messages");
 
     if (check_name_change_propagation) {
-        for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
-            for (uint16_t j = 0; j < NUM_GROUP_TOX; ++j) {
+        for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
+            for (uint32_t j = 0; j < NUM_GROUP_TOX; ++j) {
                 uint8_t name[NAMELEN];
                 tox_conference_peer_get_name(toxes[i], 0, j, name, nullptr);
                 /* Note the toxes will have been reordered */
@@ -297,14 +297,14 @@ static void run_conference_tests(Tox **toxes, State *state)
         }
     }
 
-    for (uint16_t k = NUM_GROUP_TOX; k != 0 ; --k) {
+    for (uint32_t k = NUM_GROUP_TOX; k != 0 ; --k) {
         tox_conference_delete(toxes[k - 1], 0, nullptr);
 
         for (uint8_t j = 0; j < 10 || j < NUM_GROUP_TOX; ++j) {
             iterate_all_wait(NUM_GROUP_TOX, toxes, state, ITERATION_INTERVAL);
         }
 
-        for (uint16_t i = 0; i < k - 1; ++i) {
+        for (uint32_t i = 0; i < k - 1; ++i) {
             uint32_t peer_count = tox_conference_peer_count(toxes[i], 0, nullptr);
             ck_assert_msg(peer_count == (k - 1), "\n\tBad number of group peers (post check)."
                           "\n\t\t\tExpected: %u but tox_instance(%u) only has: %u\n\n",
@@ -317,7 +317,7 @@ static void test_many_group(Tox **toxes, State *state)
 {
     const time_t test_start_time = time(nullptr);
 
-    for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
+    for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
         tox_callback_self_connection_status(toxes[i], &handle_self_connection_status);
         tox_callback_friend_connection_status(toxes[i], &handle_friend_connection_status);
         tox_callback_conference_invite(toxes[i], &handle_conference_invite);
@@ -337,21 +337,21 @@ static void test_many_group(Tox **toxes, State *state)
 
 
     printf("waiting for invitations to be made\n");
-    uint16_t invited_count = 0;
+    uint32_t invited_count = 0;
 
     do {
         iterate_all_wait(NUM_GROUP_TOX, toxes, state, ITERATION_INTERVAL);
 
         invited_count = 0;
 
-        for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
+        for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
             invited_count += state[i].invited_next;
         }
     } while (invited_count != NUM_GROUP_TOX - 1);
 
     uint64_t pregroup_clock = state[0].clock;
     printf("waiting for all toxes to be in the group\n");
-    uint16_t fully_connected_count = 0;
+    uint32_t fully_connected_count = 0;
 
     do {
         fully_connected_count = 0;
@@ -359,7 +359,7 @@ static void test_many_group(Tox **toxes, State *state)
 
         iterate_all_wait(NUM_GROUP_TOX, toxes, state, ITERATION_INTERVAL);
 
-        for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
+        for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
             Tox_Err_Conference_Peer_Query err;
             uint32_t peer_count = tox_conference_peer_count(toxes[i], 0, &err);
 
@@ -380,7 +380,7 @@ static void test_many_group(Tox **toxes, State *state)
         fflush(stdout);
     } while (fully_connected_count != NUM_GROUP_TOX);
 
-    for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
+    for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
         uint32_t peer_count = tox_conference_peer_count(toxes[i], 0, nullptr);
 
         ck_assert_msg(peer_count == NUM_GROUP_TOX, "\n\tBad number of group peers (pre check)."
